separate script for testing interface
